Restructuring API, tests passing
