+++ app/controllers/api_controller.rb
@@ -1,0 +1,5 @@
+class ApiController < ActionController::API
+  #before_action :authorize!
+
+  #ruby uuid method for access token?
+end
