--- app/controllers/request_controller.rb
@@ -1,2 +1,0 @@
-class RequestController < ActionController::API
-end
