+++ app/models/api_key_generator.rb
@@ -1,0 +1,6 @@
+module ApiKeyGenerator
+
+  def self.new_key
+    SecureRandom.hex
+  end
+end
