+++ config/routes.rb
@@ -30,12 +30,10 @@ Rails.application.routes.draw do
     resources :users, only: [:index, :show]
   end
 
-  namespace :api do
-    namespace :v1 do
-      namespace :pins do
-        get '/:id/comments', to: 'comments#index'
-        post '/:id/comments', to: 'comments#create'
-      end
+  namespace :api, defaults: {format: :json} do
+    namespace :v1 do#api/vi/comments?item_id=1
+      get '/comments', to: 'comments#index'
+      post '/comments', to: 'comments#create'
       put '/comments', to: 'comments#update'
       delete '/comments', to: 'comments#destroy'
     end
