+++ db/migrate/20161221182148_add_api_key_to_user.rb
@@ -1,0 +1,5 @@
+class AddApiKeyToUser < ActiveRecord::Migration[5.0]
+  def change
+    add_column :users, :api_key, :string
+  end
+end
